backend-microservices/shared/requirements-base.txt
Django>=4.2.29,<6.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
celery>=5.4.0
mysqlclient>=2.1.1
python-decouple>=3.8
PyMySQL>=1.0.2
djangorestframework-camel-case>=1.4.2
